+++ test/unit/replication.test.js
@@ -59,7 +59,7 @@ describe('replication.test.js', () => {
     });
     config.parallel('test pouch-sync to ensure nothing broke', () => {
         describe('positive', () => {
-            it('sync two collections over server', async function() {
+            it('sync two collections over server', async function () {
                 const server = await SpawnServer.spawn();
                 const c = await humansCollection.create(0);
                 const c2 = await humansCollection.create(0);
@@ -67,7 +67,7 @@ describe('replication.test.js', () => {
                 const pw8 = AsyncTestUtil.waitResolveable(1000);
                 c.pouch.sync(server.url, {
                     live: true
-                }).on('error', function(err) {
+                }).on('error', function (err) {
                     console.log('error:');
                     console.log(JSON.stringify(err));
                     throw new Error(err);
@@ -121,10 +121,10 @@ describe('replication.test.js', () => {
                             live: true,
                             include_docs: true
                         }), 'change')
-                    .pipe(
-                        map(ar => ar[0]),
-                        filter(e => !e.id.startsWith('_'))
-                    ).subscribe(e => e1.push(e));
+                        .pipe(
+                            map(ar => ar[0]),
+                            filter(e => !e.id.startsWith('_'))
+                        ).subscribe(e => e1.push(e));
                 const e2 = [];
                 const pouch2$ =
                     fromEvent(c2.pouch.changes({
@@ -271,6 +271,26 @@ describe('replication.test.js', () => {
         });
     });
     config.parallel('RxReplicationState', () => {
+        describe('._pouchEventEmitterObject', () => {
+            it('should be able to get the event-emitter after some time', async () => {
+                const c = await humansCollection.create(0);
+                const c2 = await humansCollection.create(10);
+                const repState = await c.sync({
+                    remote: c2,
+                    waitForLeadership: false
+                });
+
+                await AsyncTestUtil.waitUntil(
+                    () => !!repState._pouchEventEmitterObject
+                );
+                const pouchEventEmitter = repState._pouchEventEmitterObject;
+                assert.ok(pouchEventEmitter);
+                assert.equal(typeof pouchEventEmitter.on, 'function');
+
+                c.database.destroy();
+                c2.database.destroy();
+            });
+        });
         describe('change$', () => {
             it('should emit change-events', async () => {
                 const c = await humansCollection.create(0);
@@ -317,7 +337,7 @@ describe('replication.test.js', () => {
 
                 const emited = [];
                 repState.alive$.subscribe(cE => emited.push(cE));
-                
+
                 assert.equal(emited[emited.length - 1], false);
 
                 c.database.destroy();
@@ -386,7 +406,7 @@ describe('replication.test.js', () => {
                             lastEv.push.ok === true &&
                             lastEv.pull.ok === true
                         ) ret = true;
-                    } catch (e) {}
+                    } catch (e) { }
                     return ret;
                 });
                 sub.unsubscribe();
@@ -543,7 +563,7 @@ describe('replication.test.js', () => {
                 c2.database.destroy();
             });
         });
-        describe('negative', () => {});
+        describe('negative', () => { });
     });
     describe('ISSUES', () => {
         it('#630 Query cache is not being invalidated by replication', async () => {